requirements.txt
requests
cloudscraper
Pillow
